+++ app/src/main/java/woowacourse/movie/view/MainActivity.kt
@@ -13,7 +13,6 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import woowacourse.movie.R
-import woowacourse.movie.ReservationDatabase
 import woowacourse.movie.contract.MainContract
 import woowacourse.movie.data.ApplicationSettings
 import woowacourse.movie.data.reservation.LocalReservationData
@@ -21,6 +20,7 @@ import woowacourse.movie.data.reservation.ReservationData
 import woowacourse.movie.databinding.ActivityMainBinding
 import woowacourse.movie.presenter.MainPresenter
 import woowacourse.movie.view.cinema.HomeFragment
+import woowacourse.movie.view.reservation.ReservationDatabase
 import woowacourse.movie.view.reservation.ReservationHistoryFragment
 import woowacourse.movie.view.setting.SettingFragment
 import woowacourse.movie.view.util.ErrorMessage
