+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationDao.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.reservation
 
 import androidx.room.Dao
 import androidx.room.Insert
