+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.view.reservation
 
 import android.content.Context
 import androidx.room.Database
